Do not trim trailing whitespace, ws-butler is a much better approach
Deleting trailing whitespace is an annoyance when contributing to other
projects, if they're not formatted correctly.

I don't want my Emacs to change an entire source file because it wasn't
formatted correctly. `ws-butler` has the correct approach, in that it trims the
whitespace only on the lines I have touched.